[py] Only attempt to remove the profile path if one was specified

+++ py/selenium/webdriver/firefox/webdriver.py
@@ -186,12 +186,13 @@ class WebDriver(RemoteWebDriver):
             self.service.stop()
         else:
             self.binary.kill()
-        try:
-            shutil.rmtree(self.profile.path)
-            if self.profile.tempfolder is not None:
-                shutil.rmtree(self.profile.tempfolder)
-        except Exception as e:
-            print(str(e))
+        if self.profile is not None:
+            try:
+                shutil.rmtree(self.profile.path)
+                if self.profile.tempfolder is not None:
+                    shutil.rmtree(self.profile.tempfolder)
+            except Exception as e:
+                print(str(e))
 
     @property
     def firefox_profile(self):
